Remove unnecessary casts
Former-commit-id: d1ef0910455e475ee47a65fc76cbf0bafc5d8d69

+++ src/main/java/com/github/bartimaeusnek/bartworks/client/renderer/BW_EICPistonVisualizer.java
@@ -18,7 +18,7 @@ import gregtech.api.GregTech_API;
 public class BW_EICPistonVisualizer extends EntityFX {
 
     public BW_EICPistonVisualizer(World world, int x, int y, int z, int age) {
-        super(world, (double) x, ((double) y), (double) z);
+        super(world, x, (y), z);
         this.prevPosX = this.posX;
         this.prevPosY = this.posY;
         this.prevPosZ = this.posZ;
