+++ src/main/java/com/github/bartimaeusnek/bartworks/util/MurmurHash3.java
@@ -147,7 +147,7 @@ public final class MurmurHash3 {
             } else {
                 // surrogate pair
                 // int utf32 = pos < end ? (int) data.charAt(pos++) : 0;
-                int utf32 = (int) data.charAt(pos++);
+                int utf32 = data.charAt(pos++);
                 utf32 = ((code - 0xD7C0) << 10) + (utf32 & 0x3FF);
                 k2 = (0xff & (0xF0 | (utf32 >> 18))) | ((0x80 | ((utf32 >> 12) & 0x3F))) << 8
                         | ((0x80 | ((utf32 >> 6) & 0x3F))) << 16
